Add additional lints for testing

+++ crates/flipperzero/src/furi/thread.rs
@@ -260,7 +260,7 @@ pub fn set_flags(thread_id: ThreadId, flags: u32) -> Result<u32, sys::furi::Stat
     let result = unsafe { sys::furi_thread_flags_set(thread_id.0, flags) };
 
     if sys::FuriFlag(result).has_flag(sys::FuriFlagError) {
-        return Err((result as i32).into());
+        return Err(result.into());
     }
 
     Ok(result)
@@ -273,7 +273,7 @@ pub fn clear_flags(flags: u32) -> Result<u32, sys::furi::Status> {
     let result = unsafe { sys::furi_thread_flags_clear(flags) };
 
     if sys::FuriFlag(result).has_flag(sys::FuriFlagError) {
-        return Err((result as i32).into());
+        return Err(result.into());
     }
 
     Ok(result)
@@ -284,7 +284,7 @@ pub fn get_flags() -> Result<u32, sys::furi::Status> {
     let result = unsafe { sys::furi_thread_flags_get() };
 
     if sys::FuriFlag(result).has_flag(sys::FuriFlagError) {
-        return Err((result as i32).into());
+        return Err(result.into());
     }
 
     Ok(result)
@@ -302,7 +302,7 @@ pub fn wait_any_flags(
     let result = unsafe { sys::furi_thread_flags_wait(flags, options, timeout.0) };
 
     if sys::FuriFlag(result).has_flag(sys::FuriFlagError) {
-        return Err((result as i32).into());
+        return Err(result.into());
     }
 
     Ok(result)
@@ -320,7 +320,7 @@ pub fn wait_all_flags(
     let result = unsafe { sys::furi_thread_flags_wait(flags, options, timeout.0) };
 
     if sys::FuriFlag(result).has_flag(sys::FuriFlagError) {
-        return Err((result as i32).into());
+        return Err(result.into());
     }
 
     Ok(result)
